Add normalizelines to util. SqrtRem in bigint. Small improvements to EightCounters
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5006 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/atomic-util.cc
@@ -1,4 +1,4 @@
 #include "atomic-util.h"
 
 // Shared across all EightCounters instances.
-thread_local size_t internal::EightCounters::idx;
+thread_local uint8_t internal::EightCounters::idx;
